refactor: add folder's structure
